ci: update build workflow to use Gradle instead of Java for environment setup.
The changes in this commit update the build workflow configuration file `build.yml`. The modifications remove the step for setting up Java 11 environment using the Zulu distribution and instead add a step for setting up Gradle using the `gradle/gradle-build-action@v2.4.2` action. This change is made to streamline the build process and align it with the project's use of Gradle as the build system.