ci: release draft workflow updated
A change was made in the `.github/workflows/build.yml` file. The `name` of the `releaseDraft` job was modified from "Release Draft" to "Release draft". Additionally, the `needs` list was updated to include `[ build, verify, inspectCode ]` as dependencies for the job. Finally, the `permissions` section was added with the `contents` permission set to `write`.