chore: add configuration to run tests using Gradle
This commit adds a new configuration to run tests using Gradle. The configuration is named "Run Tests" and is of type GradleRunConfiguration. It includes a log file alias and path, as well as external system settings for Gradle. The configuration specifies the task to run as "check" and does not include any script parameters or VM options. The commit also sets the ExternalSystemDebugServerProcess and ExternalSystemReattachDebugProcess options to true, while setting DebugAllEnabled to false.