ci: add usage guard to Github workflow
Added a new 'guard' job to the Github workflow in build.yml file which uses 'action-usage-guard' action. This job is intended to act as a threshold guard for resource usage. It gives the ability to define a threshold and prevent further workflow execution if the threshold exceeded. The 'guard' job runs before the 'build' job, ensuring resources are kept in balance.